ADC: working

+++ common/bsp/adc.h
@@ -12,6 +12,6 @@
 //=========================== prototypes ======================================
 
 void    adc_init(void);
-uint8_t adc_read_P002(void);
+int16_t adc_read_P002(void);
 
 #endif
